requirements.txt
pycryptodome
pycryptodomex
docopt
